requirements.txt
crewai>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
graphviz>=0.20.1
pyyaml>=6.0.1 